+++ CHANGELOG.md
@@ -13,6 +13,12 @@ was added, changed, deprecated, removed, fix and security fixes.
 
 ## [Unreleased]
 
+## [0.9.2] - 2017-03-20
+### Fixed
+- Bug in SpringBootServerProvider when trying to add a new element to an immutable map
+- Bug in HK2ServiceInstance that prevented the removal and replacement services
+- Bug in ResourceInstanceBuilder where builder methods used super instead of extends
+
 ## [0.9.1] - 2017-03-19
 ### Fixed
 - Bug were deployment of artifact to sonatype failed due to [missing main artifact](https://travis-ci.org/testify-project/testify/builds/212702576)